Consolidate duplicate uses of test util `getEntityState`.
