+++ ui/apps/platform/src/utils/WorkflowState.test.js
@@ -2,65 +2,19 @@ import entityTypes from 'constants/entityTypes';
 import relationshipTypes from 'constants/relationshipTypes';
 import useCases from 'constants/useCaseTypes';
 import { searchParams, sortParams, pagingParams } from 'constants/searchParams';
+import {
+    entityId1,
+    entityId2,
+    entityId3,
+    getEntityState,
+    getListState,
+    searchParamValues,
+    sortParamValues,
+    pagingParamValues,
+} from 'test-utils/workflowUtils';
 import { getEntityTypesByRelationship, useCaseEntityMap } from './entityRelationships';
 import WorkflowEntity from './WorkflowEntity';
 import { WorkflowState } from './WorkflowState';
-
-const entityId1 = '1234';
-const entityId2 = '5678';
-const entityId3 = '1111';
-
-const searchParamValues = {
-    [searchParams.page]: {
-        sk1: 'v1',
-        sk2: 'v2',
-    },
-    [searchParams.sidePanel]: {
-        sk3: 'v3',
-        sk4: 'v4',
-    },
-};
-
-const sortParamValues = {
-    [sortParams.page]: entityTypes.CLUSTER,
-    [sortParams.sidePanel]: entityTypes.DEPLOYMENT,
-};
-
-const pagingParamValues = {
-    [pagingParams.page]: 1,
-    [pagingParams.sidePanel]: 2,
-};
-
-function getEntityState(isSidePanelOpen) {
-    const stateStack = [new WorkflowEntity(entityTypes.CLUSTER, entityId1)];
-    if (isSidePanelOpen) {
-        stateStack.push(new WorkflowEntity(entityTypes.DEPLOYMENT));
-        stateStack.push(new WorkflowEntity(entityTypes.DEPLOYMENT, entityId2));
-    }
-
-    return new WorkflowState(
-        useCases.CONFIG_MANAGEMENT,
-        stateStack,
-        searchParamValues,
-        sortParamValues,
-        pagingParamValues
-    );
-}
-
-function getListState(isSidePanelOpen) {
-    const stateStack = [new WorkflowEntity(entityTypes.CLUSTER)];
-    if (isSidePanelOpen) {
-        stateStack.push(new WorkflowEntity(entityTypes.CLUSTER, entityId1));
-    }
-
-    return new WorkflowState(
-        useCases.CONFIG_MANAGEMENT,
-        stateStack,
-        searchParamValues,
-        sortParamValues,
-        pagingParamValues
-    );
-}
 
 describe('WorkflowState', () => {
     it('clears current state on current use case', () => {
