+++ ui/apps/platform/src/utils/workflowUtils.test.ts
@@ -1,8 +1,6 @@
 import entityTypes from 'constants/entityTypes';
-import { searchParams, sortParams, pagingParams } from 'constants/searchParams';
-import useCases from 'constants/useCaseTypes';
-import WorkflowEntity from 'utils/WorkflowEntity';
-import { WorkflowState } from 'utils/WorkflowState';
+
+import { getEntityState } from '../test-utils/workflowUtils';
 
 // system under test (SUT)
 import { createOptions, getOption, shouldUseOriginalCase } from './workflowUtils';
@@ -75,43 +73,3 @@ describe('workflowUtils', () => {
         });
     });
 });
-
-const entityId1 = '1234';
-const entityId2 = '5678';
-
-const searchParamValues = {
-    [searchParams.page]: {
-        sk1: 'v1',
-        sk2: 'v2',
-    },
-    [searchParams.sidePanel]: {
-        sk3: 'v3',
-        sk4: 'v4',
-    },
-};
-
-const sortParamValues = {
-    [sortParams.page]: entityTypes.CLUSTER,
-    [sortParams.sidePanel]: entityTypes.DEPLOYMENT,
-};
-
-const pagingParamValues = {
-    [pagingParams.page]: 1,
-    [pagingParams.sidePanel]: 2,
-};
-
-function getEntityState(isSidePanelOpen) {
-    const stateStack = [new WorkflowEntity(entityTypes.CLUSTER, entityId1)];
-    if (isSidePanelOpen) {
-        stateStack.push(new WorkflowEntity(entityTypes.DEPLOYMENT));
-        stateStack.push(new WorkflowEntity(entityTypes.DEPLOYMENT, entityId2));
-    }
-
-    return new WorkflowState(
-        useCases.CONFIG_MANAGEMENT,
-        stateStack,
-        searchParamValues,
-        sortParamValues,
-        pagingParamValues
-    );
-}
